Moving more functionality to the individual packages
