+++ src/mantle/application/class-application-exception.php
@@ -5,7 +5,7 @@
  * @package Mantle
  */
 
-namespace Mantle\Framework;
+namespace Mantle\Application;
 
 /**
  * Application Exception
