+++ src/mantle/events/class-event-service-provider.php
@@ -1,0 +1,28 @@
+<?php
+/**
+ * Event_Service_Provider class file.
+ *
+ * @package Mantle
+ */
+
+namespace Mantle\Events;
+
+use Mantle\Support\Service_Provider;
+
+/**
+ * Event Service Provider
+ *
+ * Registered the internal events dispatcher. The application's event service
+ * provider extends {@see \Mantle\Framework\Providers\Event_Service_Provider}.
+ */
+class Event_Service_Provider extends Service_Provider {
+	/**
+	 * Register any application services.
+	 */
+	public function register() {
+		$this->app->singleton_if(
+			'events',
+			fn( $app ) => new Dispatcher( $app ),
+		);
+	}
+}
